+++ fs/src/fs/org/jnode/fs/hfsplus/SuperBlock.java
@@ -209,7 +209,7 @@ public class SuperBlock extends HfsPlusObject {
     // Getters/setters
 
     public final int getMagic() {
-        return BigEndian.getInt16(data, 0);
+        return BigEndian.getUInt16(data, 0);
     }
 
     public final void setMagic(final int value) {
@@ -218,7 +218,7 @@ public class SuperBlock extends HfsPlusObject {
 
     //
     public final int getVersion() {
-        return BigEndian.getInt16(data, 2);
+        return BigEndian.getUInt16(data, 2);
     }
 
     public final void setVersion(final int value) {
@@ -226,18 +226,18 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getAttributes() {
+    public final long getAttributes() {
 
-        return BigEndian.getInt32(data, 4);
+        return BigEndian.getUInt32(data, 4);
     }
 
     public final void setAttribute(final int attributeMaskBit) {
-        BigEndian.setInt32(data, 4, getAttributes() | (1 << attributeMaskBit));
+        BigEndian.setInt32(data, 4, (int)(getAttributes() | (1 << attributeMaskBit)));
     }
 
     //
-    public final int getLastMountedVersion() {
-        return BigEndian.getInt32(data, 8);
+    public final long getLastMountedVersion() {
+        return BigEndian.getUInt32(data, 8);
     }
 
     public final void setLastMountedVersion(final int value) {
@@ -245,8 +245,8 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getJournalInfoBlock() {
-        return BigEndian.getInt32(data, 12);
+    public final long getJournalInfoBlock() {
+        return BigEndian.getUInt32(data, 12);
     }
 
     public final void setJournalInfoBlock(final long value) {
@@ -287,8 +287,8 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getFileCount() {
-        return BigEndian.getInt32(data, 32);
+    public final long getFileCount() {
+        return BigEndian.getUInt32(data, 32);
     }
 
     public final void setFileCount(final int value) {
@@ -296,12 +296,12 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getFolderCount() {
-        return BigEndian.getInt32(data, 36);
+    public final long getFolderCount() {
+        return BigEndian.getUInt32(data, 36);
     }
 
-    public final void setFolderCount(final int value) {
-        BigEndian.setInt32(data, 36, value);
+    public final void setFolderCount(final long value) {
+        BigEndian.setInt32(data, 36, (int) value);
     }
 
     //
@@ -314,8 +314,8 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getTotalBlocks() {
-        return BigEndian.getInt32(data, 44);
+    public final long getTotalBlocks() {
+        return BigEndian.getUInt32(data, 44);
     }
 
     public final void setTotalBlocks(final int value) {
@@ -323,17 +323,17 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     //
-    public final int getFreeBlocks() {
-        return BigEndian.getInt32(data, 48);
+    public final long getFreeBlocks() {
+        return BigEndian.getUInt32(data, 48);
     }
 
-    public final void setFreeBlocks(final int value) {
-        BigEndian.setInt32(data, 48, value);
+    public final void setFreeBlocks(final long value) {
+        BigEndian.setInt32(data, 48, (int) value);
     }
 
     //
-    public final int getNextAllocation() {
-        return BigEndian.getInt32(data, 52);
+    public final long getNextAllocation() {
+        return BigEndian.getUInt32(data, 52);
     }
 
     public final void setNextAllocation(final int value) {
@@ -341,31 +341,31 @@ public class SuperBlock extends HfsPlusObject {
     }
 
     public final long getRsrcClumpSize() {
-        return BigEndian.getInt32(data, 56);
+        return BigEndian.getUInt32(data, 56);
     }
 
     public final void setRsrcClumpSize(final int value) {
         BigEndian.setInt32(data, 56, value);
     }
 
-    public final int getDataClumpSize() {
-        return BigEndian.getInt32(data, 60);
+    public final long getDataClumpSize() {
+        return BigEndian.getUInt32(data, 60);
     }
 
     public final void setDataClumpSize(final int value) {
         BigEndian.setInt32(data, 60, value);
     }
 
-    public final int getNextCatalogId() {
-        return BigEndian.getInt32(data, 64);
+    public final long getNextCatalogId() {
+        return BigEndian.getUInt32(data, 64);
     }
 
-    public final void setNextCatalogId(final int value) {
-        BigEndian.setInt32(data, 64, value);
+    public final void setNextCatalogId(final long value) {
+        BigEndian.setInt32(data, 64, (int) value);
     }
 
-    public final int getWriteCount() {
-        return BigEndian.getInt32(data, 68);
+    public final long getWriteCount() {
+        return BigEndian.getUInt32(data, 68);
     }
 
     public final void setWriteCount(final int value) {
