+++ fs/src/fs/org/jnode/fs/hfsplus/catalog/CatalogFolder.java
@@ -32,13 +32,13 @@ public class CatalogFolder {
     public static final int CATALOG_FOLDER_SIZE = 88;
 
     private int recordType;
-    private int valence;
+    private long valence;
     private CatalogNodeId folderId;
-    private int createDate;
-    private int contentModDate;
-    private int attrModDate;
-    private int accessDate;
-    private int backupDate;
+    private long createDate;
+    private long contentModDate;
+    private long attrModDate;
+    private long accessDate;
+    private long backupDate;
     private HfsPlusBSDInfo permissions;
 
     /**
@@ -49,13 +49,14 @@ public class CatalogFolder {
         byte[] data = new byte[88];
         System.arraycopy(src, 0, data, 0, CATALOG_FOLDER_SIZE);
         recordType = BigEndian.getInt16(data, 0);
-        valence = BigEndian.getInt32(data, 4);
+        // UInt16 flags - offset 2
+        valence = BigEndian.getUInt32(data, 4);
         folderId = new CatalogNodeId(data, 8);
-        createDate = BigEndian.getInt32(data, 12);
-        contentModDate = BigEndian.getInt32(data, 16);
-        attrModDate = BigEndian.getInt32(data, 20);
-        accessDate = BigEndian.getInt32(data, 24);
-        backupDate = BigEndian.getInt32(data, 28);
+        createDate = BigEndian.getUInt32(data, 12);
+        contentModDate = BigEndian.getUInt32(data, 16);
+        attrModDate = BigEndian.getUInt32(data, 20);
+        accessDate = BigEndian.getUInt32(data, 24);
+        backupDate = BigEndian.getUInt32(data, 28);
         permissions = new HfsPlusBSDInfo(data, 32);
     }
 
@@ -81,11 +82,11 @@ public class CatalogFolder {
     public byte[] getBytes() {
         byte[] data = new byte[88];
         BigEndian.setInt16(data, 0, recordType);
-        BigEndian.setInt32(data, 4, valence);
+        BigEndian.setInt32(data, 4, (int) valence);
         System.arraycopy(folderId.getBytes(), 0, data, 8, folderId.getBytes().length);
-        BigEndian.setInt32(data, 12, createDate);
-        BigEndian.setInt32(data, 16, contentModDate);
-        BigEndian.setInt32(data, 20, attrModDate);
+        BigEndian.setInt32(data, 12, (int) createDate);
+        BigEndian.setInt32(data, 16, (int) contentModDate);
+        BigEndian.setInt32(data, 20, (int) attrModDate);
         return data;
     }
 
@@ -111,7 +112,7 @@ public class CatalogFolder {
         return recordType;
     }
 
-    public int getValence() {
+    public long getValence() {
         return valence;
     }
 
@@ -147,7 +148,7 @@ public class CatalogFolder {
         this.recordType = recordType;
     }
 
-    public void setValence(int valence) {
+    public void setValence(long valence) {
         this.valence = valence;
     }
 
@@ -156,15 +157,15 @@ public class CatalogFolder {
     }
 
     public void setCreateDate(long createDate) {
-        this.createDate = (int) HfsUtils.getDate(createDate / 1000L, true);
+        this.createDate = HfsUtils.getDate(createDate / 1000L, true);
     }
 
     public void setContentModDate(long contentModDate) {
-        this.contentModDate = (int) HfsUtils.getDate(contentModDate / 1000L, true);
+        this.contentModDate = HfsUtils.getDate(contentModDate / 1000L, true);
     }
 
     public void setAttrModDate(long attrModDate) {
-        this.attrModDate = (int) HfsUtils.getDate(attrModDate / 1000L, true);
+        this.attrModDate = HfsUtils.getDate(attrModDate / 1000L, true);
     }
     
     public void incrementValence(){
